renaming _MembraneBaseData _ReverseOsmosisBaseData; changing 0DRO defaults to align with 1DRO
