+++ watertap/unit_models/tests/test_reverse_osmosis_0D.py
@@ -69,9 +69,9 @@ def test_config():
     assert m.fs.unit.config.property_package is \
            m.fs.properties
     assert m.fs.unit.config.concentration_polarization_type == \
-           ConcentrationPolarizationType.none
+           ConcentrationPolarizationType.calculated
     assert m.fs.unit.config.mass_transfer_coefficient == \
-           MassTransferCoefficient.none
+           MassTransferCoefficient.calculated
     assert m.fs.unit.config.pressure_change_type == \
            PressureChangeType.fixed_per_stage
 
@@ -95,7 +95,8 @@ def test_option_concentration_polarization_type_fixed():
     m.fs.unit = ReverseOsmosis0D(default={
         "property_package": m.fs.properties,
         "has_pressure_change": True,
-        "concentration_polarization_type": ConcentrationPolarizationType.fixed})
+        "concentration_polarization_type": ConcentrationPolarizationType.fixed,
+        "mass_transfer_coefficient": MassTransferCoefficient.none})
 
     assert m.fs.unit.config.concentration_polarization_type == \
            ConcentrationPolarizationType.fixed
@@ -182,7 +183,8 @@ class TestReverseOsmosis():
         m.fs.unit = ReverseOsmosis0D(default={
             "property_package": m.fs.properties,
             "has_pressure_change": True,
-            "concentration_polarization_type": ConcentrationPolarizationType.fixed})
+            "concentration_polarization_type": ConcentrationPolarizationType.fixed,
+            "mass_transfer_coefficient": MassTransferCoefficient.none})
 
         # fully specify system
         feed_flow_mass = 1
